Refactor overall structure
One of the main drivers behind this refactor is
to reduce the class-focus, i.e. stop using classes
where there is no real need. For example, the
Game class is completely gone after this refactor,
without any major impact.

The refactor seems to result in a more readable
code base.